481 Readjusted structure for tests user and login
